Style fixes for cast_unittest to build
Relanding r226264 that includes a bunch of style fixes to cast code.

R=hubbe@chromium.org

Review URL: https://codereview.chromium.org/25782002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@226538 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 248f7d04b66fcda8500685455e67f51c623410c7

+++ media/cast/rtp_sender/rtp_sender.cc
@@ -83,7 +83,7 @@ void RtpSender::ResendPackets(
           // Set a unique incremental sequence number for every packet.
           UpdateSequenceNumber(&packet);
           // Set the size as correspond to each frame.
-          transport_->ResendPacket(packet, packets.size());
+          transport_->ResendPacket(packet, static_cast<int>(packets.size()));
           ++packet_id;
         }
       } while (success);
@@ -101,7 +101,7 @@ void RtpSender::ResendPackets(
               << packet_id << " size: " << packet.size();
           UpdateSequenceNumber(&packet);
           // Set the size as correspond to each frame.
-          transport_->ResendPacket(packet, packets.size());
+          transport_->ResendPacket(packet, static_cast<int>(packets.size()));
         } else {
           VLOG(1) << "Failed to resend " << static_cast<int>(frame_id) << ":"
               << packet_id;
